*** RELEASE 0.5.1 ***
Summary of Changes:
 * Now the array can be returned as a "simple" array.
 * Use "__METHOD__" in exceptions to ensure the proper method is being named.

NOTE: when a simple array is returned, there is no way to expose the attributes 
of a given tag (through the output array).  Should still be able to use the 
method "get_attribute()" to retrieve that information.

NOTE2: version 0.5 was never released due to a mistake on my part.  This is a 
very minor enhancement to what that release was.

/xmlParserClass.php:
	* MAIN:::
		-- new private var/property, $makeSimpleTree = FALSE.
	* get_tree():
		-- ARG CHANGE: NEW ARG: #1 ($simpleTree=FALSE)
		-- when $simpleTree evaluates as TRUE, it sets the internal property 
		"makeSimpleTree"
	* build_tag():
		-- when the internal "makeSimpleTree" is set, it builds the tag very 
		plainly, any extra array elements. This avoids extra processing when the 
		code doesn't care about attributes.