Rewrite cs_phpxmlCreator{} to remove limitations & reserved words.
NOTE::: After the rewrite, it will store a list of paths with values to
indicate where it should "live" in the final XML. The cs_arrayToPath{} class
is not used until the point that the XML string is actually created.  Paths
will be stored with numeric separators for simple identification of multiples
and such (i.e. in "/ROOT/0/PATH/0/TO/3/HEAVEN", the "TO" tag has 3 other tags
of the same name at the same level, making at least 4).

cs_phpxmlCreator.class.php:
	* THE FOLLOWING METHODS THROW EXCEPTIONS UNTIL UPDATED:::
		-- add_attribute()
		-- create_tag()
		-- reconstruct_path()
		-- set_tag_as_multiple()
		-- create_path()
		-- add_tag_multiple()
		-- rename_root_element()
		-- get_data()
		-- load_xmlparser_data()
		-- remove_path()
		-- path_exists()
	* MAIN:::
		-- updated comments regarding retrieval of data within a specific tag 
		(where that tag is a multiple).
		-- changed example XML to be upper-case, to avoid confusion.
		-- added code that would (theoretically) create the example XML.
	* add_tag():
		-- TODO about checking numeric indexes within path
		-- use paths array after fixing given path
		-- no more debug_print() or exit statements.
		-- throws an exception if there's no length to the path after fixing.
	* fix_path():
		-- changed so the path is given with numeric separators between tags 
		to indicate explicit paths
		-- handles paths that have only some numeric separators
		-- throws exceptions as needed if the path is invalid